refactor(vector): move field options under the schema module
Follow the convention of facet_options.rs/numeric_options.rs and keep
the vector field-type surface (VectorOptions, Metric, VectorDType) in
schema/vector_options.rs. The VectorElement storage trait stays in the
vector module.
