+++ src/schema/field_type.rs
@@ -14,12 +14,11 @@ use crate::schema::bytes_options::BytesOptions;
 use crate::schema::facet_options::FacetOptions;
 use crate::schema::{
     DateOptions, Facet, IndexRecordOption, JsonObjectOptions, NumericOptions, OwnedValue,
-    TextFieldIndexing, TextOptions,
+    TextFieldIndexing, TextOptions, VectorOptions,
 };
 use crate::time::format_description::well_known::Rfc3339;
 use crate::time::OffsetDateTime;
 use crate::tokenizer::PreTokenizedString;
-use crate::vector::VectorOptions;
 use crate::DateTime;
 
 /// Possible error that may occur while parsing a field value
