+++ src/schema/mod.rs
@@ -129,6 +129,7 @@ mod json_object_options;
 mod named_field_document;
 mod numeric_options;
 mod text_options;
+mod vector_options;
 
 use columnar::ColumnType;
 
@@ -150,6 +151,7 @@ pub use self::numeric_options::NumericOptions;
 pub use self::schema::{Schema, SchemaBuilder};
 pub use self::term::{Term, ValueBytes};
 pub use self::text_options::{TextFieldIndexing, TextOptions, STRING, TEXT};
+pub use self::vector_options::{Metric, VectorDType, VectorOptions};
 
 /// Validator for a potential `field_name`.
 /// Returns true if the name can be use for a field name.
