+++ src/vector/mod.rs
@@ -1,3 +1,54 @@
-mod options;
+use crate::schema::VectorDType;
 
-pub use options::{Metric, VectorDType, VectorElement, VectorOptions};
+/// A vector element type with the primitives needed by the storage
+/// layer and the distance kernels.
+///
+/// Implemented for the element types supported by [`VectorDType`]. The
+/// `DTYPE` associated constant lets callers reject mismatches between
+/// the declared schema dtype and the type passed at runtime. The
+/// arithmetic methods (`squared_diff`, `product`) return `f32` so that
+/// kernels can use a uniform accumulator type across dtypes.
+pub trait VectorElement: Copy + Send + Sync + 'static {
+    const DTYPE: VectorDType;
+    const SIZE_BYTES: usize;
+
+    fn encode_le(&self, buf: &mut Vec<u8>);
+
+    /// Decode one element from its little-endian byte representation.
+    /// `bytes.len()` must be `SIZE_BYTES`.
+    fn decode_le(bytes: &[u8]) -> Self;
+
+    /// `(a - b)^2` promoted to `f32` for accumulator-friendly distance
+    /// computation. For `f32` this is the obvious arithmetic; for
+    /// quantized types it may promote through a wider integer first.
+    fn squared_diff(a: Self, b: Self) -> f32;
+
+    /// `a * b` promoted to `f32`. Same rationale as `squared_diff`.
+    fn product(a: Self, b: Self) -> f32;
+}
+
+impl VectorElement for f32 {
+    const DTYPE: VectorDType = VectorDType::F32;
+    const SIZE_BYTES: usize = 4;
+
+    #[inline(always)]
+    fn encode_le(&self, buf: &mut Vec<u8>) {
+        buf.extend_from_slice(&self.to_le_bytes());
+    }
+
+    #[inline(always)]
+    fn decode_le(bytes: &[u8]) -> Self {
+        f32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]])
+    }
+
+    #[inline(always)]
+    fn squared_diff(a: Self, b: Self) -> f32 {
+        let d = a - b;
+        d * d
+    }
+
+    #[inline(always)]
+    fn product(a: Self, b: Self) -> f32 {
+        a * b
+    }
+}
