--- src/vector/options.rs
@@ -1,117 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum VectorDType {
-    F32,
-}
-
-impl VectorDType {
-    pub fn size_bytes(self) -> usize {
-        match self {
-            VectorDType::F32 => 4,
-        }
-    }
-}
-
-/// Distance / similarity metric used when ranking vector field values.
-///
-/// All metrics are presented to callers in a "higher is better" orientation.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Metric {
-    L2,
-    Cosine,
-    Dot,
-}
-
-/// A vector element type with the primitives needed by the storage
-/// layer and the distance kernels.
-///
-/// Implemented for the element types supported by [`VectorDType`]. The
-/// `DTYPE` associated constant lets callers reject mismatches between
-/// the declared schema dtype and the type passed at runtime. The
-/// arithmetic methods (`squared_diff`, `product`) return `f32` so that
-/// kernels can use a uniform accumulator type across dtypes.
-pub trait VectorElement: Copy + Send + Sync + 'static {
-    const DTYPE: VectorDType;
-    const SIZE_BYTES: usize;
-
-    fn encode_le(&self, buf: &mut Vec<u8>);
-
-    /// Decode one element from its little-endian byte representation.
-    /// `bytes.len()` must be `SIZE_BYTES`.
-    fn decode_le(bytes: &[u8]) -> Self;
-
-    /// `(a - b)^2` promoted to `f32` for accumulator-friendly distance
-    /// computation. For `f32` this is the obvious arithmetic; for
-    /// quantized types it may promote through a wider integer first.
-    fn squared_diff(a: Self, b: Self) -> f32;
-
-    /// `a * b` promoted to `f32`. Same rationale as `squared_diff`.
-    fn product(a: Self, b: Self) -> f32;
-}
-
-impl VectorElement for f32 {
-    const DTYPE: VectorDType = VectorDType::F32;
-    const SIZE_BYTES: usize = 4;
-
-    #[inline(always)]
-    fn encode_le(&self, buf: &mut Vec<u8>) {
-        buf.extend_from_slice(&self.to_le_bytes());
-    }
-
-    #[inline(always)]
-    fn decode_le(bytes: &[u8]) -> Self {
-        f32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]])
-    }
-
-    #[inline(always)]
-    fn squared_diff(a: Self, b: Self) -> f32 {
-        let d = a - b;
-        d * d
-    }
-
-    #[inline(always)]
-    fn product(a: Self, b: Self) -> f32 {
-        a * b
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VectorOptions {
-    dim: usize,
-    dtype: VectorDType,
-    metric: Metric,
-}
-
-impl VectorOptions {
-    pub fn new(dim: usize, metric: Metric) -> VectorOptions {
-        VectorOptions {
-            dim,
-            dtype: VectorDType::F32,
-            metric,
-        }
-    }
-
-    pub fn dim(&self) -> usize {
-        self.dim
-    }
-
-    pub fn dtype(&self) -> VectorDType {
-        self.dtype
-    }
-
-    pub fn metric(&self) -> Metric {
-        self.metric
-    }
-
-    pub fn with_dtype(mut self, dtype: VectorDType) -> VectorOptions {
-        self.dtype = dtype;
-        self
-    }
-
-    pub fn bytes_per_vector(&self) -> usize {
-        self.dim * self.dtype.size_bytes()
-    }
-}
